Move DapRequest and it's extractors to their own modules
